+++ R/hash.R
@@ -37,6 +37,8 @@
 #' When you [`tidy()`][tidy.recipe()] this step, a tibble with columns
 #' `terms` (the columns that is selected)  is returned.
 #' 
+#' @template case-weights-not-supported
+#' 
 #' @references
 #' Weinberger, K, A Dasgupta, J Langford, A Smola, and J Attenberg. 2009.
 #'  "Feature Hashing for Large Scale Multitask Learning." In Proceedings of the
@@ -197,7 +199,8 @@ make_hash_tbl <- function(ind, nms) {
 
 #' @export
 bake.step_feature_hash <- function(object, new_data, ...) {
-
+  check_new_data(names(object$columns), object, new_data)
+  
   # If no terms were selected
   if (length(object$columns) == 0) {
     return(new_data)