+++ tests/testthat/test_discretize_cart.R
@@ -171,6 +171,19 @@ test_that("tidy method", {
   )
 })
 
+test_that("bake method errors when needed non-standard role columns are missing", {
+  rec <- recipe(class ~ ., data = sim_tr_cls) %>%
+    step_discretize_cart(x, z, outcome = "class") %>%
+    update_role(x, new_role = "potato") %>%
+    update_role_requirements(role = "potato", bake = FALSE)
+  
+  expect_warning(
+    rec_trained <- prep(rec, training = sim_tr_cls, verbose = FALSE)
+  )
+ 
+  expect_error(bake(rec_trained, new_data = sim_tr_cls[, -1]),
+               class = "new_data_missing_column")
+})
 
 # ------------------------------------------------------------------------------
 
@@ -180,7 +193,7 @@ test_that("printing", {
     step_discretize_cart(all_predictors(), outcome = "class")
 
   expect_snapshot(cart_rec)
-  expect_snapshot(prep(cart_rec, verbose = TRUE))
+  expect_snapshot(prep(cart_rec))
 })
 
 
@@ -199,4 +212,39 @@ test_that("empty selections", {
     bake(rec, new_data = NULL),
     ad_data %>% select(Genotype, tau, Class)
   )
+})
+
+test_that("case weights step functions", {
+  sim_tr_cls_cw <- sim_tr_cls %>%
+    mutate(weight = importance_weights(rep(0:1, each = 500)))
+  
+  sim_tr_reg_cw <- sim_tr_reg %>%
+    mutate(weight = importance_weights(rep(0:1, each = 500)))
+  
+  mod_cw <- rpart(y ~ x, data = sim_tr_reg, weights = rep(0:1, each = 500))
+  best_split_cw <- unname(mod_cw$splits[, "index"])
+
+  # Classification
+  expect_snapshot({
+    cart_rec <-
+      recipe(class ~ ., data = sim_tr_cls_cw) %>%
+      step_discretize_cart(all_predictors(), outcome = "class") %>%
+      prep()
+  })
+  
+  expect_equal(names(cart_rec$steps[[1]]$rules), "x")
+  expect_equal(cart_rec$steps[[1]]$rules$x, best_split_cw)
+  
+  # Regression
+  expect_snapshot({
+    cart_rec <-
+      recipe(y ~ ., data = sim_tr_reg_cw) %>%
+      step_discretize_cart(all_predictors(), outcome = "y") %>%
+      prep()
+  })
+  
+  expect_equal(names(cart_rec$steps[[1]]$rules), c("x", "z"))
+  expect_equal(cart_rec$steps[[1]]$rules$x, best_split_cw)
+  
+  expect_snapshot(cart_rec)
 })