+++ tests/testthat/test_no_pooling.R
@@ -249,12 +249,23 @@ test_that("bad args", {
   )
 })
 
+test_that("bake method errors when needed non-standard role columns are missing", {
+  rec <- recipe(x2 ~ ., data = ex_dat) %>%
+    step_lencode_glm(x3, outcome = vars(x2)) %>%
+    update_role(x3, new_role = "potato") %>%
+    update_role_requirements(role = "potato", bake = FALSE)
+  
+  rec_trained <- prep(rec, training = ex_dat, verbose = FALSE)
+  
+  expect_error(bake(rec_trained, new_data = ex_dat[, -3]),
+               class = "new_data_missing_column")
+})
 
 test_that("printing", {
   print_test <- recipe(x2 ~ ., data = ex_dat_ch) %>%
     step_lencode_glm(x3, outcome = vars(x2))
   expect_snapshot(print_test)
-  expect_snapshot(prep(print_test, training = ex_dat_ch, verbose = TRUE))
+  expect_snapshot(prep(print_test))
 })
 
 
@@ -273,4 +284,31 @@ test_that("empty selections", {
     bake(rec, new_data = NULL),
     ad_data %>% select(Genotype, tau, Class)
   )
+})
+
+# ------------------------------------------------------------------------------
+
+test_that("case weights", {
+  wts_int <- rep(c(0, 1), times = c(100, 400))
+  
+  ex_dat_cw <- ex_dat %>%
+    mutate(wts = importance_weights(wts_int))
+  
+  class_test <- recipe(x2 ~ ., data = ex_dat_cw) %>%
+    step_lencode_glm(x3, outcome = vars(x2), id = "id") %>%
+    prep(training = ex_dat_cw, retain = TRUE)
+  
+  ref_mod <- glm(
+    x2 ~ 0 + x3,
+    data = ex_dat_cw,
+    family = binomial,
+    na.action = na.omit, weights = ex_dat_cw$wts
+  )
+  
+  expect_equal(
+    -unname(coef(ref_mod)),
+    slice_head(class_test$steps[[1]]$mapping$x3, n = -1)$..value
+  )
+  
+  expect_snapshot(class_test)
 })